fix migrations to work with old and new instances, also with sqlite tests and mariadb/mysql

+++ database/migrations/2023_11_27_000000_drop_delay_column_in_hafas_trips.php
@@ -7,9 +7,11 @@ use Illuminate\Support\Facades\Schema;
 return new class extends Migration
 {
     public function up(): void {
-        Schema::table('hafas_trips', static function(Blueprint $table) {
-            $table->dropColumn('delay');
-        });
+        if (Schema::hasColumn('hafas_trips', 'delay')) {
+            Schema::table('hafas_trips', static function(Blueprint $table) {
+                $table->dropColumn('delay');
+            });
+        }
     }
 
     public function down(): void {
